% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variables.R
\name{add_variables}
\alias{add_variables}
\title{Add Variables}
\usage{
add_variables(.data)
}
\arguments{
\item{.data}{A data.frame with matches data
generated by brasileirao package.}
}
\value{
A data.frame with added variables.
}
\description{
Add home_score, away_score, home_points,
away_points, home_abbr and away_abbr variables to
a matches dataset.
}
\examples{
add_variables(matches)

}
