% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{badges}
\alias{badges}
\title{Badge file from each team in the matches data framse}
\format{
A data frame with 2 variables:
\describe{
\item{team}{The team name.}
\item{badge}{The badge filename that can be accessed with
\code{system.file("badges/", package = "brasileirao")}}
}
}
\source{
\url{https://github.com/williamorim/brasileirao}
}
\usage{
badges
}
\description{
A dataset containing the filename from each team in the matches data frame.
}
\keyword{datasets}
