% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_names}
\alias{fix_names}
\title{Standardize names}
\usage{
fix_names(teams)
}
\arguments{
\item{teams}{Character vector with the name of the teams you want
to standardize.}
}
\value{
Character vector with the standardized team names.
}
\description{
Standardize the name of the teams to what is used nowadays.
}
\examples{
fix_names("Prudente")

}
