% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variables.R
\name{get_abbr_name}
\alias{get_abbr_name}
\title{Get Abbreviated Team Name}
\usage{
get_abbr_name(team)
}
\arguments{
\item{team}{The full name of the team.}
}
\value{
The abbreviated name of the team.
}
\description{
This function retrieves the abbreviated name of a team from the Brazilian
football league.
}
\examples{
team_name <- "São Paulo"
get_abbr_name(team_name)

}
