% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variables.R
\name{get_away_score}
\alias{get_away_score}
\title{Get Away Team Score}
\usage{
get_away_score(scores)
}
\arguments{
\item{scores}{A character vector containing scores.}
}
\value{
A numeric vector of the away team scores extracted from the
input scores.
}
\description{
This function extracts the away team score from a vector of scores.
}
\examples{
scores <- c("3x2", "1x0", "2x2")
get_away_score(scores)

}
