% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_badge}
\alias{get_badge}
\title{Get team badge}
\usage{
get_badge(team)
}
\arguments{
\item{team}{A Brazilian team. See \code{badges} to see the available
badges in this packages. The function will remove accents, uppercase
letters and whitespaces.}
}
\value{
A string with the path to the badge image file.
}
\description{
Return the path to a badge image file.
}
\examples{
get_badge("sao paulo")
get_badge("sao-paulo")
}
