% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matches}
\alias{matches}
\title{Brazilian National Soccer League matches}
\format{
A data frame with 5 variables:
\describe{
\item{season}{The year of national league edition.}
\item{date}{The match date.}
\item{home}{The home match team.}
\item{score}{The score of the match in the format home x away.}
\item{away}{The away match team.}
}
}
\source{
\url{https://github.com/williamorim/brasileirao}
}
\usage{
matches
}
\description{
A dataset containing the matches Brazilian National Soccer League
(Brasileirão) matches from 2003 to 2020.
}
\keyword{datasets}
