% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teams}
\alias{teams}
\title{Teams of the \code{matches} dataset}
\format{
A data frame with 5 variables:
\describe{
\item{team}{Team name.}
\item{abbr}{Team name abbreviation.}
\item{state}{The state where the club is based.}
\item{color1}{The most frequent color in the team badge.}
\item{color2}{The second more frequent color in the team badge.}
}
}
\source{
\url{https://github.com/williamorim/brasileirao}
}
\usage{
teams
}
\description{
A dataset containing more information about each
one of the teams in the \code{matches} dataset.
}
\keyword{datasets}
